requirements.txt
# Core ML libraries (Python 3.12 compatible)
numpy==2.2.0
scipy==1.14.1
scikit-learn==1.6.0
pandas==2.2.3

# FastAPI and web server (latest)
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database and cache
pymongo==4.10.1
redis==5.2.1

# Utilities
python-dotenv==1.0.1